Fix additive coloring
Previously, I used a hovering counter to calculate the
brightness of the cell coloring. This created an unexpected
behavior. When changing the brightness parameter of one pass,
if the counter*brightness was smaller than the current
color brightness, then the cell color became lighter.

To fix this I changed the hovering counter by a color property with
initial value 255 and then I subtract the brightness*255 on each
pass wrapping the result to the interval [0,255].

This issue inspired allowing negative values for the brightness
parameter to remove color on each pass.